Data Structures for Moving and Spatial Object
The main aim of this paper is to provide a Data Structure to support
queries and perform operations on time dependent geometric data stored
as Spatio-temporal(Moving) data. The data of moving object is stored
with the time dependent position of object and manipulated. To come up
with a data structure we first come up with a data model involving
complex spatial structures and operations, which can be plugged into
DBMS, which support query language. A concept of sliced representation
is introduced, it represents temporal development of moving objects as a
set of units, temporal development within each such unit can be
represented by simple linear function of time. Then an algorithm
describing the use and implementation of the data structure and its
complexity is discussed. Other data structures which solve the
challenges faced by moving objects are also discussed.